code/lecture05/requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
torch
numpy
PyPDF2
gTTS
pyttsx3 